refactor: Add assembler for post and sheet to replace service convert
