+++ src/main/java/run/halo/app/controller/content/api/PostController.java
@@ -22,10 +22,11 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.HtmlUtils;
 import run.halo.app.cache.lock.CacheLock;
 import run.halo.app.cache.lock.CacheParam;
+import run.halo.app.controller.content.auth.PostAuthentication;
+import run.halo.app.exception.ForbiddenException;
 import run.halo.app.exception.NotFoundException;
 import run.halo.app.model.dto.BaseCommentDTO;
 import run.halo.app.model.dto.post.BasePostSimpleDTO;
-import run.halo.app.model.entity.Content;
 import run.halo.app.model.entity.Post;
 import run.halo.app.model.entity.PostComment;
 import run.halo.app.model.enums.CommentStatus;
@@ -40,6 +41,7 @@ import run.halo.app.model.vo.PostListVO;
 import run.halo.app.service.OptionService;
 import run.halo.app.service.PostCommentService;
 import run.halo.app.service.PostService;
+import run.halo.app.service.assembler.PostRenderAssembler;
 
 /**
  * Content post controller.
@@ -59,18 +61,25 @@ public class PostController {
 
     private final OptionService optionService;
 
+    private final PostRenderAssembler postRenderAssembler;
+
+    private final PostAuthentication postAuthentication;
+
     public PostController(PostService postService,
         PostCommentService postCommentService,
-        OptionService optionService) {
+        OptionService optionService, PostRenderAssembler postRenderAssembler,
+        PostAuthentication postAuthentication) {
         this.postService = postService;
         this.postCommentService = postCommentService;
         this.optionService = optionService;
+        this.postRenderAssembler = postRenderAssembler;
+        this.postAuthentication = postAuthentication;
     }
 
     //CS304 issue for https://github.com/halo-dev/halo/issues/1351
 
     /**
-     * Enable users search published articles with keywords
+     * Enable users search published articles with keywords.
      *
      * @param pageable store the priority of the sort algorithm
      * @param keyword search articles with keyword
@@ -88,7 +97,7 @@ public class PostController {
         postQuery.setCategoryId(categoryId);
         postQuery.setStatuses(Set.of(PostStatus.PUBLISHED));
         Page<Post> postPage = postService.pageBy(postQuery, pageable);
-        return postService.convertToListVo(postPage, true);
+        return postRenderAssembler.convertToListVo(postPage);
     }
 
     @PostMapping(value = "search")
@@ -96,7 +105,7 @@ public class PostController {
     public Page<BasePostSimpleDTO> pageBy(@RequestParam(value = "keyword") String keyword,
         @PageableDefault(sort = "createTime", direction = DESC) Pageable pageable) {
         Page<Post> postPage = postService.pageBy(keyword, pageable);
-        return postService.convertToSimple(postPage);
+        return postRenderAssembler.convertToSimple(postPage);
     }
 
     @GetMapping("{postId:\\d+}")
@@ -107,8 +116,10 @@ public class PostController {
         @RequestParam(value = "sourceDisabled", required = false, defaultValue = "false")
             Boolean sourceDisabled) {
         Post post = postService.getById(postId);
-        post.setContent(Content.PatchedContent.of(postService.getContentById(postId)));
-        PostDetailVO postDetailVO = postService.convertToDetailVo(post);
+
+        checkAuthenticate(postId);
+
+        PostDetailVO postDetailVO = postRenderAssembler.convertToDetailVo(post);
 
         if (formatDisabled) {
             // Clear the format content
@@ -133,8 +144,10 @@ public class PostController {
         @RequestParam(value = "sourceDisabled", required = false, defaultValue = "false")
             Boolean sourceDisabled) {
         Post post = postService.getBySlug(slug);
-        post.setContent(Content.PatchedContent.of(postService.getContentById(post.getId())));
-        PostDetailVO postDetailVO = postService.convertToDetailVo(post);
+
+        checkAuthenticate(post.getId());
+
+        PostDetailVO postDetailVO = postRenderAssembler.convertToDetailVo(post);
 
         if (formatDisabled) {
             // Clear the format content
@@ -157,9 +170,8 @@ public class PostController {
         Post post = postService.getById(postId);
         Post prevPost =
             postService.getPrevPost(post).orElseThrow(() -> new NotFoundException("查询不到该文章的信息"));
-        prevPost.setContent(
-            Content.PatchedContent.of(postService.getContentById(prevPost.getId())));
-        return postService.convertToDetailVo(prevPost);
+        checkAuthenticate(prevPost.getId());
+        return postRenderAssembler.convertToDetailVo(prevPost);
     }
 
     @GetMapping("{postId:\\d+}/next")
@@ -168,15 +180,15 @@ public class PostController {
         Post post = postService.getById(postId);
         Post nextPost =
             postService.getNextPost(post).orElseThrow(() -> new NotFoundException("查询不到该文章的信息"));
-        nextPost.setContent(
-            Content.PatchedContent.of(postService.getContentById(nextPost.getId())));
-        return postService.convertToDetailVo(nextPost);
+        checkAuthenticate(nextPost.getId());
+        return postRenderAssembler.convertToDetailVo(nextPost);
     }
 
     @GetMapping("{postId:\\d+}/comments/top_view")
     public Page<CommentWithHasChildrenVO> listTopComments(@PathVariable("postId") Integer postId,
         @RequestParam(name = "page", required = false, defaultValue = "0") int page,
         @SortDefault(sort = "createTime", direction = DESC) Sort sort) {
+        checkAuthenticate(postId);
         return postCommentService.pageTopCommentsBy(postId, CommentStatus.PUBLISHED,
             PageRequest.of(page, optionService.getCommentPageSize(), sort));
     }
@@ -185,6 +197,7 @@ public class PostController {
     public List<BaseCommentDTO> listChildrenBy(@PathVariable("postId") Integer postId,
         @PathVariable("commentParentId") Long commentParentId,
         @SortDefault(sort = "createTime", direction = DESC) Sort sort) {
+        checkAuthenticate(postId);
         // Find all children comments
         List<PostComment> postComments = postCommentService
             .listChildrenBy(postId, commentParentId, CommentStatus.PUBLISHED, sort);
@@ -198,6 +211,7 @@ public class PostController {
     public Page<BaseCommentVO> listCommentsTree(@PathVariable("postId") Integer postId,
         @RequestParam(name = "page", required = false, defaultValue = "0") int page,
         @SortDefault(sort = "createTime", direction = DESC) Sort sort) {
+        checkAuthenticate(postId);
         return postCommentService
             .pageVosBy(postId, PageRequest.of(page, optionService.getCommentPageSize(), sort));
     }
@@ -207,6 +221,7 @@ public class PostController {
     public Page<BaseCommentWithParentVO> listComments(@PathVariable("postId") Integer postId,
         @RequestParam(name = "page", required = false, defaultValue = "0") int page,
         @SortDefault(sort = "createTime", direction = DESC) Sort sort) {
+        checkAuthenticate(postId);
         return postCommentService.pageWithParentVoBy(postId,
             PageRequest.of(page, optionService.getCommentPageSize(), sort));
     }
@@ -215,6 +230,7 @@ public class PostController {
     @ApiOperation("Comments a post")
     @CacheLock(autoDelete = false, traceRequest = true)
     public BaseCommentDTO comment(@RequestBody PostCommentParam postCommentParam) {
+        checkAuthenticate(postCommentParam.getPostId());
         postCommentService.validateCommentBlackListStatus();
 
         // Escape content
@@ -227,6 +243,13 @@ public class PostController {
     @ApiOperation("Likes a post")
     @CacheLock(autoDelete = false, traceRequest = true)
     public void like(@PathVariable("postId") @CacheParam Integer postId) {
+        checkAuthenticate(postId);
         postService.increaseLike(postId);
+    }
+
+    private void checkAuthenticate(Integer postId) {
+        if (!postAuthentication.isAuthenticated(postId)) {
+            throw new ForbiddenException("您没有该分类的访问权限");
+        }
     }
 }
