+++ src/main/java/run/halo/app/controller/content/auth/PostAuthentication.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Component;
 import run.halo.app.cache.AbstractStringCacheStore;
 import run.halo.app.model.entity.Post;
 import run.halo.app.model.enums.EncryptTypeEnum;
+import run.halo.app.service.CategoryService;
+import run.halo.app.service.PostCategoryService;
 import run.halo.app.service.PostService;
 
 /**
@@ -18,11 +20,17 @@ import run.halo.app.service.PostService;
 public class PostAuthentication implements ContentAuthentication {
 
     private final PostService postService;
+    private final CategoryService categoryService;
+    private final PostCategoryService postCategoryService;
     private final AbstractStringCacheStore cacheStore;
 
     public PostAuthentication(PostService postService,
+        CategoryService categoryService,
+        PostCategoryService postCategoryService,
         AbstractStringCacheStore cacheStore) {
         this.postService = postService;
+        this.categoryService = categoryService;
+        this.postCategoryService = postCategoryService;
         this.cacheStore = cacheStore;
     }
 
@@ -34,8 +42,12 @@ public class PostAuthentication implements ContentAuthentication {
     @Override
     public boolean isAuthenticated(Integer postId) {
         Post post = postService.getById(postId);
-        if (post.getPassword() == null) {
-            return true;
+        if (StringUtils.isBlank(post.getPassword())) {
+            boolean categoryEncrypted = postCategoryService.listByPostId(postId).stream()
+                .anyMatch(postCategory -> categoryService.isPrivate(postCategory.getCategoryId()));
+            if (!categoryEncrypted) {
+                return true;
+            }
         }
 
         String sessionId = getSessionId();
