+++ src/main/java/run/halo/app/service/impl/PostServiceImpl.java
@@ -1,18 +1,14 @@
 package run.halo.app.service.impl;
 
 import static org.springframework.data.domain.Sort.Direction.DESC;
-import static run.halo.app.model.support.HaloConst.URL_SEPARATOR;
 
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -29,7 +25,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
@@ -38,8 +33,6 @@ import run.halo.app.event.logger.LogEvent;
 import run.halo.app.event.post.PostUpdatedEvent;
 import run.halo.app.event.post.PostVisitEvent;
 import run.halo.app.exception.NotFoundException;
-import run.halo.app.model.dto.post.BasePostMinimalDTO;
-import run.halo.app.model.dto.post.BasePostSimpleDTO;
 import run.halo.app.model.entity.Category;
 import run.halo.app.model.entity.Content;
 import run.halo.app.model.entity.Content.PatchedContent;
@@ -49,9 +42,7 @@ import run.halo.app.model.entity.PostComment;
 import run.halo.app.model.entity.PostMeta;
 import run.halo.app.model.entity.PostTag;
 import run.halo.app.model.entity.Tag;
-import run.halo.app.model.enums.CommentStatus;
 import run.halo.app.model.enums.LogType;
-import run.halo.app.model.enums.PostPermalinkType;
 import run.halo.app.model.enums.PostStatus;
 import run.halo.app.model.params.PostParam;
 import run.halo.app.model.params.PostQuery;
@@ -59,7 +50,6 @@ import run.halo.app.model.properties.PostProperties;
 import run.halo.app.model.vo.ArchiveMonthVO;
 import run.halo.app.model.vo.ArchiveYearVO;
 import run.halo.app.model.vo.PostDetailVO;
-import run.halo.app.model.vo.PostListVO;
 import run.halo.app.model.vo.PostMarkdownVO;
 import run.halo.app.repository.PostRepository;
 import run.halo.app.repository.base.BasePostRepository;
@@ -73,6 +63,7 @@ import run.halo.app.service.PostMetaService;
 import run.halo.app.service.PostService;
 import run.halo.app.service.PostTagService;
 import run.halo.app.service.TagService;
+import run.halo.app.service.assembler.PostAssembler;
 import run.halo.app.utils.DateUtils;
 import run.halo.app.utils.HaloUtils;
 import run.halo.app.utils.MarkdownUtils;
@@ -93,6 +84,8 @@ import run.halo.app.utils.SlugUtils;
 @Slf4j
 @Service
 public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostService {
+
+    private final PostAssembler postAssembler;
 
     private final PostRepository postRepository;
 
@@ -119,7 +112,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
     private final ApplicationContext applicationContext;
 
     public PostServiceImpl(BasePostRepository<Post> basePostRepository,
-        OptionService optionService,
+        PostAssembler postAssembler, OptionService optionService,
         PostRepository postRepository,
         TagService tagService,
         CategoryService categoryService,
@@ -132,6 +125,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         ContentPatchLogService contentPatchLogService,
         ApplicationContext applicationContext) {
         super(basePostRepository, optionService, contentService, contentPatchLogService);
+        this.postAssembler = postAssembler;
         this.postRepository = postRepository;
         this.tagService = tagService;
         this.categoryService = categoryService;
@@ -315,7 +309,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         List<Post> posts = postRepository
             .findAllByStatus(PostStatus.PUBLISHED, Sort.by(DESC, "createTime"));
 
-        return convertToYearArchives(posts);
+        return postAssembler.convertToYearArchives(posts);
     }
 
     @Override
@@ -324,67 +318,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         List<Post> posts = postRepository
             .findAllByStatus(PostStatus.PUBLISHED, Sort.by(DESC, "createTime"));
 
-        return convertToMonthArchives(posts);
-    }
-
-    @Override
-    public List<ArchiveYearVO> convertToYearArchives(List<Post> posts) {
-        Map<Integer, List<Post>> yearPostMap = new HashMap<>(8);
-
-        posts.forEach(post -> {
-            Calendar calendar = DateUtils.convertTo(post.getCreateTime());
-            yearPostMap.computeIfAbsent(calendar.get(Calendar.YEAR), year -> new LinkedList<>())
-                .add(post);
-        });
-
-        List<ArchiveYearVO> archives = new LinkedList<>();
-
-        yearPostMap.forEach((year, postList) -> {
-            // Build archive
-            ArchiveYearVO archive = new ArchiveYearVO();
-            archive.setYear(year);
-            archive.setPosts(convertToListVo(postList));
-
-            // Add archive
-            archives.add(archive);
-        });
-
-        // Sort this list
-        archives.sort(new ArchiveYearVO.ArchiveComparator());
-
-        return archives;
-    }
-
-    @Override
-    public List<ArchiveMonthVO> convertToMonthArchives(List<Post> posts) {
-
-        Map<Integer, Map<Integer, List<Post>>> yearMonthPostMap = new HashMap<>(8);
-
-        posts.forEach(post -> {
-            Calendar calendar = DateUtils.convertTo(post.getCreateTime());
-
-            yearMonthPostMap.computeIfAbsent(calendar.get(Calendar.YEAR), year -> new HashMap<>())
-                .computeIfAbsent(calendar.get(Calendar.MONTH) + 1,
-                    month -> new LinkedList<>())
-                .add(post);
-        });
-
-        List<ArchiveMonthVO> archives = new LinkedList<>();
-
-        yearMonthPostMap.forEach((year, monthPostMap) ->
-            monthPostMap.forEach((month, postList) -> {
-                ArchiveMonthVO archive = new ArchiveMonthVO();
-                archive.setYear(year);
-                archive.setMonth(month);
-                archive.setPosts(convertToListVo(postList));
-
-                archives.add(archive);
-            }));
-
-        // Sort this list
-        archives.sort(new ArchiveMonthVO.ArchiveComparator());
-
-        return archives;
+        return postAssembler.convertToMonthArchives(posts);
     }
 
     @Override
@@ -556,30 +490,6 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
     }
 
     @Override
-    public PostDetailVO convertToDetailVo(Post post) {
-        return convertToDetailVo(post, false);
-    }
-
-    @Override
-    public PostDetailVO convertToDetailVo(Post post, boolean queryEncryptCategory) {
-        // List tags
-        List<Tag> tags = postTagService.listTagsBy(post.getId());
-        // List categories
-        List<Category> categories = postCategoryService
-            .listCategoriesBy(post.getId());
-        // List metas
-        List<PostMeta> metas = postMetaService.listBy(post.getId());
-        // Convert to detail vo
-        return convertTo(post, tags, categories, metas);
-    }
-
-    @Override
-    public Page<PostDetailVO> convertToDetailVo(Page<Post> postPage) {
-        Assert.notNull(postPage, "Post page must not be null");
-        return postPage.map(this::convertToDetailVo);
-    }
-
-    @Override
     public Post removeById(Integer postId) {
         Assert.notNull(postId, "Post id must not be null");
 
@@ -614,217 +524,6 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
             deletedPost.getTitle()));
 
         return deletedPost;
-    }
-
-    @Override
-    public Page<PostListVO> convertToListVo(Page<Post> postPage) {
-        return convertToListVo(postPage, false);
-    }
-
-    @Override
-    public Page<PostListVO> convertToListVo(Page<Post> postPage, boolean queryEncryptCategory) {
-        Assert.notNull(postPage, "Post page must not be null");
-
-        List<Post> posts = postPage.getContent();
-
-        Set<Integer> postIds = ServiceUtils.fetchProperty(posts, Post::getId);
-
-        // Get tag list map
-        Map<Integer, List<Tag>> tagListMap = postTagService.listTagListMapBy(postIds);
-
-        // Get category list map
-        Map<Integer, List<Category>> categoryListMap = postCategoryService
-            .listCategoryListMap(postIds);
-
-        // Get comment count
-        Map<Integer, Long> commentCountMap = postCommentService.countByStatusAndPostIds(
-            CommentStatus.PUBLISHED, postIds);
-
-        // Get post meta list map
-        Map<Integer, List<PostMeta>> postMetaListMap = postMetaService.listPostMetaAsMap(postIds);
-
-        return postPage.map(post -> {
-            PostListVO postListVO = new PostListVO().convertFrom(post);
-
-            generateAndSetSummaryIfAbsent(post, postListVO);
-
-            Optional.ofNullable(tagListMap.get(post.getId())).orElseGet(LinkedList::new);
-
-            // Set tags
-            postListVO.setTags(Optional.ofNullable(tagListMap.get(post.getId()))
-                .orElseGet(LinkedList::new)
-                .stream()
-                .filter(Objects::nonNull)
-                .map(tagService::convertTo)
-                .collect(Collectors.toList()));
-
-            // Set categories
-            postListVO.setCategories(Optional.ofNullable(categoryListMap.get(post.getId()))
-                .orElseGet(LinkedList::new)
-                .stream()
-                .filter(Objects::nonNull)
-                .map(categoryService::convertTo)
-                .collect(Collectors.toList()));
-
-            // Set post metas
-            List<PostMeta> metas = Optional.ofNullable(postMetaListMap.get(post.getId()))
-                .orElseGet(LinkedList::new);
-            postListVO.setMetas(postMetaService.convertToMap(metas));
-
-            // Set comment count
-            postListVO.setCommentCount(commentCountMap.getOrDefault(post.getId(), 0L));
-
-            postListVO.setFullPath(buildFullPath(post));
-
-            // Post currently drafting in process
-            Boolean isInProcess = postContentService.draftingInProgress(post.getId());
-            postListVO.setInProgress(isInProcess);
-
-            return postListVO;
-        });
-    }
-
-    @Override
-    public List<PostListVO> convertToListVo(List<Post> posts) {
-        Assert.notNull(posts, "Post page must not be null");
-
-        Set<Integer> postIds = ServiceUtils.fetchProperty(posts, Post::getId);
-
-        // Get tag list map
-        Map<Integer, List<Tag>> tagListMap = postTagService.listTagListMapBy(postIds);
-
-        // Get category list map
-        Map<Integer, List<Category>> categoryListMap = postCategoryService
-            .listCategoryListMap(postIds);
-
-        // Get comment count
-        Map<Integer, Long> commentCountMap =
-            postCommentService.countByStatusAndPostIds(CommentStatus.PUBLISHED, postIds);
-
-        // Get post meta list map
-        Map<Integer, List<PostMeta>> postMetaListMap = postMetaService.listPostMetaAsMap(postIds);
-
-        return posts.stream().map(post -> {
-            PostListVO postListVO = new PostListVO().convertFrom(post);
-
-            generateAndSetSummaryIfAbsent(post, postListVO);
-
-            Optional.ofNullable(tagListMap.get(post.getId())).orElseGet(LinkedList::new);
-
-            // Set tags
-            postListVO.setTags(Optional.ofNullable(tagListMap.get(post.getId()))
-                .orElseGet(LinkedList::new)
-                .stream()
-                .filter(Objects::nonNull)
-                .map(tagService::convertTo)
-                .collect(Collectors.toList()));
-
-            // Set categories
-            postListVO.setCategories(Optional.ofNullable(categoryListMap.get(post.getId()))
-                .orElseGet(LinkedList::new)
-                .stream()
-                .filter(Objects::nonNull)
-                .map(categoryService::convertTo)
-                .collect(Collectors.toList()));
-
-            // Set post metas
-            List<PostMeta> metas = Optional.ofNullable(postMetaListMap.get(post.getId()))
-                .orElseGet(LinkedList::new);
-            postListVO.setMetas(postMetaService.convertToMap(metas));
-
-            // Set comment count
-            postListVO.setCommentCount(commentCountMap.getOrDefault(post.getId(), 0L));
-
-            postListVO.setFullPath(buildFullPath(post));
-
-            return postListVO;
-        }).collect(Collectors.toList());
-    }
-
-    @Override
-    public BasePostMinimalDTO convertToMinimal(Post post) {
-        Assert.notNull(post, "Post must not be null");
-        BasePostMinimalDTO basePostMinimalDTO = new BasePostMinimalDTO().convertFrom(post);
-
-        basePostMinimalDTO.setFullPath(buildFullPath(post));
-
-        return basePostMinimalDTO;
-    }
-
-    @Override
-    public List<BasePostMinimalDTO> convertToMinimal(List<Post> posts) {
-        if (CollectionUtils.isEmpty(posts)) {
-            return Collections.emptyList();
-        }
-
-        return posts.stream()
-            .map(this::convertToMinimal)
-            .collect(Collectors.toList());
-    }
-
-    @Override
-    public BasePostSimpleDTO convertToSimple(Post post) {
-        Assert.notNull(post, "Post must not be null");
-
-        BasePostSimpleDTO basePostSimpleDTO = new BasePostSimpleDTO().convertFrom(post);
-
-        // Set summary
-        generateAndSetSummaryIfAbsent(post, basePostSimpleDTO);
-
-        basePostSimpleDTO.setFullPath(buildFullPath(post));
-
-        return basePostSimpleDTO;
-    }
-
-    /**
-     * Converts to post detail vo.
-     *
-     * @param post post must not be null
-     * @param tags tags
-     * @param categories categories
-     * @param postMetaList postMetaList
-     * @return post detail vo
-     */
-    @NonNull
-    private PostDetailVO convertTo(@NonNull Post post, @Nullable List<Tag> tags,
-        @Nullable List<Category> categories, List<PostMeta> postMetaList) {
-        Assert.notNull(post, "Post must not be null");
-
-        // Convert to base detail vo
-        PostDetailVO postDetailVO = new PostDetailVO().convertFrom(post);
-        generateAndSetSummaryIfAbsent(post, postDetailVO);
-
-        // Extract ids
-        Set<Integer> tagIds = ServiceUtils.fetchProperty(tags, Tag::getId);
-        Set<Integer> categoryIds = ServiceUtils.fetchProperty(categories, Category::getId);
-        Set<Long> metaIds = ServiceUtils.fetchProperty(postMetaList, PostMeta::getId);
-
-        // Get post tag ids
-        postDetailVO.setTagIds(tagIds);
-        postDetailVO.setTags(tagService.convertTo(tags));
-
-        // Get post category ids
-        postDetailVO.setCategoryIds(categoryIds);
-        postDetailVO.setCategories(categoryService.convertTo(categories));
-
-        // Get post meta ids
-        postDetailVO.setMetaIds(metaIds);
-        postDetailVO.setMetas(postMetaService.convertTo(postMetaList));
-
-        postDetailVO.setCommentCount(postCommentService.countByStatusAndPostId(
-            CommentStatus.PUBLISHED, post.getId()));
-
-        postDetailVO.setFullPath(buildFullPath(post));
-
-        PatchedContent postContent = post.getContent();
-        postDetailVO.setContent(postContent.getContent());
-        postDetailVO.setOriginalContent(postContent.getOriginalContent());
-
-        // Post currently drafting in process
-        Boolean inProgress = postContentService.draftingInProgress(post.getId());
-        postDetailVO.setInProgress(inProgress);
-
-        return postDetailVO;
     }
 
     /**
@@ -926,7 +625,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         post.setContent(
             postContentPatchLogService.getPatchedContentById(postContent.getHeadPatchLogId()));
         // Convert to post detail vo
-        return convertTo(post, tags, categories, postMetaList);
+        return postAssembler.convertTo(post, tags, categories, postMetaList);
     }
 
     @Override
@@ -960,7 +659,7 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         frontMatter.append("updated: ").append(post.getUpdateTime()).append("\n");
 
         //set fullPath
-        frontMatter.append("url: ").append(buildFullPath(post)).append("\n");
+        frontMatter.append("url: ").append(postAssembler.buildFullPath(post)).append("\n");
 
         //set category
         List<Category> categories = postCategoryService.listCategoriesBy(post.getId());
@@ -997,67 +696,5 @@ public class PostServiceImpl extends BasePostServiceImpl<Post> implements PostSe
         postMarkdownVO.setTitle(post.getTitle());
         postMarkdownVO.setSlug(post.getSlug());
         return postMarkdownVO;
-    }
-
-    private String buildFullPath(Post post) {
-
-        PostPermalinkType permalinkType = optionService.getPostPermalinkType();
-
-        String pathSuffix = optionService.getPathSuffix();
-
-        String archivesPrefix = optionService.getArchivesPrefix();
-
-        int month = DateUtils.month(post.getCreateTime()) + 1;
-
-        String monthString = month < 10 ? "0" + month : String.valueOf(month);
-
-        int day = DateUtils.dayOfMonth(post.getCreateTime());
-
-        String dayString = day < 10 ? "0" + day : String.valueOf(day);
-
-        StringBuilder fullPath = new StringBuilder();
-
-        if (optionService.isEnabledAbsolutePath()) {
-            fullPath.append(optionService.getBlogBaseUrl());
-        }
-
-        fullPath.append(URL_SEPARATOR);
-
-        if (permalinkType.equals(PostPermalinkType.DEFAULT)) {
-            fullPath.append(archivesPrefix)
-                .append(URL_SEPARATOR)
-                .append(post.getSlug())
-                .append(pathSuffix);
-        } else if (permalinkType.equals(PostPermalinkType.ID)) {
-            fullPath.append("?p=")
-                .append(post.getId());
-        } else if (permalinkType.equals(PostPermalinkType.DATE)) {
-            fullPath.append(DateUtils.year(post.getCreateTime()))
-                .append(URL_SEPARATOR)
-                .append(monthString)
-                .append(URL_SEPARATOR)
-                .append(post.getSlug())
-                .append(pathSuffix);
-        } else if (permalinkType.equals(PostPermalinkType.DAY)) {
-            fullPath.append(DateUtils.year(post.getCreateTime()))
-                .append(URL_SEPARATOR)
-                .append(monthString)
-                .append(URL_SEPARATOR)
-                .append(dayString)
-                .append(URL_SEPARATOR)
-                .append(post.getSlug())
-                .append(pathSuffix);
-        } else if (permalinkType.equals(PostPermalinkType.YEAR)) {
-            fullPath.append(DateUtils.year(post.getCreateTime()))
-                .append(URL_SEPARATOR)
-                .append(post.getSlug())
-                .append(pathSuffix);
-        } else if (permalinkType.equals(PostPermalinkType.ID_SLUG)) {
-            fullPath.append(archivesPrefix)
-                .append(URL_SEPARATOR)
-                .append(post.getId())
-                .append(pathSuffix);
-        }
-        return fullPath.toString();
     }
 }
